requirements.txt
pyyaml>=6.0
pathlib2>=2.3.0; python_version<"3.4" 